conversa/requirements.txt
asgiref==3.7.2
attrs==23.1.0
black==23.10.1
channels==4.0.0
click==8.1.7
Django==4.2.7
djangorestframework==3.14.0
drf-spectacular==0.26.5
exceptiongroup==1.1.3
flake8==6.1.0
h11==0.14.0
httptools==0.6.1
idna==3.4
inflection==0.5.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
Pillow==10.1.0
platformdirs==3.11.0
pycodestyle==2.11.1
pyflakes==3.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.12.0
sniffio==1.3.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
uritemplate==4.1.1
uvicorn==0.24.0.post1
# uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
